refactor(base-driver): Lazy loading for SDKs (improve memory usage + speedup start)
